extra tests

+++ api/cas-server-core-api-monitor/src/main/java/org/apereo/cas/monitor/CacheStatistics.java
@@ -40,7 +40,7 @@ public interface CacheStatistics {
      *
      * @return Percent of space/capacity free.
      */
-    default int getPercentFree() {
+    default long getPercentFree() {
         return 0;
     }
